переделать логику визуализации лайков
